requirements.txt
streamlit==1.37.0
playwright==1.45.0
pillow==10.4.0
pandas==2.2.2
openpyxl==3.1.5
XlsxWriter==3.2.0
python-docx==1.1.2
